updating surrogate_optimizer example
