+++ meta_automl/data_preparation/surrogate_dataset/__init__.py
@@ -1,9 +1,11 @@
 from .data import GraphDataset, PairDataset, SingleDataset
-from .dataset_generate import KnowledgeBaseToDataset
+from .dataset_generate import KnowledgeBaseToDataset, dataset_from_id_without_data_loading, dataset_from_id_with_data_loading
 
 __all__ = [
     "GraphDataset",
     "SingleDataset",
     "PairDataset",
     "KnowledgeBaseToDataset",
+    "dataset_from_id_without_data_loading",
+    "dataset_from_id_with_data_loading"
 ]
