+++ meta_automl/data_preparation/surrogate_dataset/dataset_generate.py
@@ -5,6 +5,7 @@ import json
 import os
 import pathlib
 import pickle
+import pandas as pd
 
 from typing import Any, Dict, List, Optional, Tuple, Union
 
@@ -17,6 +18,40 @@ from meta_automl.data_preparation.model import Model
 from meta_automl.data_preparation.models_loaders import KnowledgeBaseModelsLoader
 from meta_automl.data_preparation.pipeline_features_extractors import FEDOTPipelineFeaturesExtractor
 
+from meta_automl.data_preparation.dataset import (CustomDataset,
+                                                  DataNotFoundError,
+                                                  DatasetData, DatasetIDType)
+
+def dataset_from_id_without_data_loading(dataset_id: DatasetIDType) -> CustomDataset:
+    """ Creates the CustomDataset object without loading the data. Use if your don't need the models
+    to load the datasets data into memory, or if you have loaded the cache manually. """
+    return CustomDataset(dataset_id)
+
+def dataset_from_id_with_data_loading(dataset_id: DatasetIDType) -> CustomDataset:
+    """ Load dataset from '//10.9.14.114/calc/Nikitin/datasets/' into the project cache directory.
+    As a result, every model of the knowledge base will have its data available by
+    `model.dataset.get_data()`.
+    """
+    dataset = CustomDataset(dataset_id)
+    try:
+        dataset.get_data()
+    except DataNotFoundError:
+        data_root = '//10.9.14.114/calc/Nikitin/datasets/'
+        dataset_name, fold_num = dataset_id[:-2], dataset_id[-1]
+        data_path = f'{dataset_name}_fold{fold_num}.npy'
+        data_x = []
+        for path_prefix in ('train_', 'test_'):
+            data_x.append(np.load(data_root + path_prefix + data_path))
+        data_y = []
+        for path_prefix in ('trainy_', 'testy_'):
+            data_y.append(np.load(data_root + path_prefix + data_path))
+        data_x = np.concatenate(data_x)
+        data_y = np.concatenate(data_y)
+        data_x = pd.DataFrame(data_x)
+        data_y = pd.DataFrame(data_y)
+        dataset_data = DatasetData(data_x, data_y)
+        dataset.dump_data(dataset_data)
+    return dataset
 
 
 def calc_pipeline_hash(pl: Pipeline) -> str:
