+++ scripts/generate_training_dataset.py
@@ -4,7 +4,6 @@ import sys
 import json
 
 import numpy as np
-import pandas as pd
 from typing import Dict
 
 sys.path.append(os.getcwd())
@@ -17,43 +16,8 @@ from meta_automl.data_preparation.feature_preprocessors import FeaturesPreproces
 from meta_automl.data_preparation.meta_features_extractors import (OpenMLDatasetMetaFeaturesExtractor,
                                                                    PymfeExtractor)
 from meta_automl.data_preparation.surrogate_dataset.dataset_generate import KnowledgeBaseToDataset
+from meta_automl.surrogate.data_pipeline_surrogate import  get_extractor_params
 
-
-def dataset_from_id_without_data_loading(dataset_id: DatasetIDType) -> CustomDataset:
-    """ Creates the CustomDataset object without loading the data. Use if your don't need the models
-    to load the datasets data into memory, or if you have loaded the cache manually. """
-    return CustomDataset(dataset_id)
-
-def dataset_from_id_with_data_loading(dataset_id: DatasetIDType) -> CustomDataset:
-    """ Load dataset from '//10.9.14.114/calc/Nikitin/datasets/' into the project cache directory.
-    As a result, every model of the knowledge base will have its data available by
-    `model.dataset.get_data()`.
-    """
-    dataset = CustomDataset(dataset_id)
-    try:
-        dataset.get_data()
-    except DataNotFoundError:
-        data_root = '//10.9.14.114/calc/Nikitin/datasets/'
-        dataset_name, fold_num = dataset_id[:-2], dataset_id[-1]
-        data_path = f'{dataset_name}_fold{fold_num}.npy'
-        data_x = []
-        for path_prefix in ('train_', 'test_'):
-            data_x.append(np.load(data_root + path_prefix + data_path))
-        data_y = []
-        for path_prefix in ('trainy_', 'testy_'):
-            data_y.append(np.load(data_root + path_prefix + data_path))
-        data_x = np.concatenate(data_x)
-        data_y = np.concatenate(data_y)
-        data_x = pd.DataFrame(data_x)
-        data_y = pd.DataFrame(data_y)
-        dataset_data = DatasetData(data_x, data_y)
-        dataset.dump_data(dataset_data)
-    return dataset
-
-def get_extractor_params(filename: str) -> Dict[str, str]:
-    with open(filename) as f:
-        extractor_params = json.load(f)
-    return extractor_params
 
 def main():
     # datasets_loader_builder = lambda: CustomDatasetsLoader(dataset_from_id_func=dataset_from_id_with_data_loading)
